changed prover.sh to print the right error message when no certificate is given
git-svn-id: https://scm.gforge.inria.fr/svn/parsifal/code/proverif@9988 ae10dad5-531c-0410-a33e-f483eea19440

+++ src/prover.sh
@@ -13,10 +13,14 @@ rm -f $RUNNER_TARGET
 
 echo -e "module runner.\\n" >> $RUNNER_TARGET
 
-for (( i=0;i<$ELEMENTS;i++)); do
-      echo -e "accumulate ${ARGS[${i}]}." >> $RUNNER_TARGET
-done
+if [ $ELEMENTS = 0 ]
+  then echo "Error: at least one certificate must be given!"
+  else
+  for (( i=0;i<$ELEMENTS;i++)); do
+        echo -e "accumulate ${ARGS[${i}]}." >> $RUNNER_TARGET
+  done
 
-echo $(cat $RUNNER_BASE) >> $RUNNER_TARGET
+  echo $(cat $RUNNER_BASE) >> $RUNNER_TARGET
 
-make>/dev/null && $CMD && rm -f $RUNNER_TARGET
+  make>/dev/null && $CMD && rm -f $RUNNER_TARGET
+fi
